cityProject/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
openai==1.3.5
pillow==10.1.0
python-slugify==8.0.1
email-validator==2.1.0
boto3==1.29.0
gunicorn==21.2.0